requirements.txt
telethon~=1.36.0
python-dotenv~=1.0.1
sqlalchemy~=2.0.31
psycopg2